refactor(FR-2444): extract ClusterModeFormItems and add hideClusterFormItems option
- Extract cluster_mode and cluster_size form items into standalone
  ClusterModeFormItems component
- Add hideClusterFormItems prop to ResourceAllocationFormItems
- In Service Launcher, hide cluster in resource allocation and show
  it in the advanced settings card instead
